+++ spec/filters/geoip_spec.rb
@@ -1,31 +1,10 @@
+# encoding: utf-8
 require "logstash/devutils/rspec/spec_helper"
 require "logstash/filters/geoip"
 
-ASNDB = ::Dir.glob(::File.expand_path("../../vendor/", ::File.dirname(__FILE__))+"/GeoIPASNum*.dat").first
-CITYDB = ::Dir.glob(::File.expand_path("../../vendor/", ::File.dirname(__FILE__))+"/GeoLiteCity*.dat").first
+CITYDB = ::Dir.glob(::File.expand_path("../../vendor/", ::File.dirname(__FILE__))+"/GeoLite2-City.mmdb").first
 
 describe LogStash::Filters::GeoIP do
-
-  describe "ASN db" do
-    config <<-CONFIG
-      filter {
-        geoip {
-          source => "ip"
-          database => "#{ASNDB}"
-        }
-      }
-    CONFIG
-
-    sample("ip" => "1.1.1.1") do
-      insist { subject["geoip"]["asn"] } == "Google Inc."
-    end
-
-    # avoid crashing on unsupported IPv6 addresses
-    # see https://github.com/logstash-plugins/logstash-filter-geoip/issues/21
-    sample("ip" => "2a02:8071:aa1:c700:7984:22fc:c8e6:f6ff") do
-      reject { subject }.include?("geoip")
-    end
-  end
 
   describe "defaults" do
     config <<-CONFIG
@@ -42,16 +21,16 @@ describe LogStash::Filters::GeoIP do
 
       expected_fields = %w(ip country_code2 country_code3 country_name
                            continent_code region_name city_name postal_code
-                           latitude longitude dma_code area_code timezone
+                           latitude longitude dma_code timezone
                            location )
       expected_fields.each do |f|
-        insist { subject["geoip"] }.include?(f)
+        insist { subject.get("geoip") }.include?(f)
       end
     end
 
     sample("ip" => "127.0.0.1") do
       # assume geoip fails on localhost lookups
-      reject { subject }.include?("geoip")
+      expect(subject.get("geoip")).to eq({})
     end
   end
 
@@ -74,24 +53,56 @@ describe LogStash::Filters::GeoIP do
 
         expected_fields = %w(ip country_code2 country_code3 country_name
                              continent_code region_name city_name postal_code
-                             latitude longitude dma_code area_code timezone
+                             latitude longitude dma_code timezone
                              location )
         expected_fields.each do |f|
-          expect(subject["src_ip"]).to include(f)
+          expect(subject.get("src_ip")).to include(f)
         end
       end
 
       sample("ip" => "127.0.0.1") do
         # assume geoip fails on localhost lookups
-        expect(subject).not_to include("src_ip")
+        expect(subject.get("src_ip")).to eq({})
       end
     end
 
     context "when specifying add_tag" do
       sample("ip" => "8.8.8.8") do
-        expect(subject["tags"]).to include("done")
+        expect(subject.get("tags")).to include("done")
       end
     end
+  end
+
+  describe "source is derived from target" do
+    subject(:event) { LogStash::Event.new("target" => { "ip" => "173.9.34.107" } ) }
+    let(:plugin) {
+      LogStash::Filters::GeoIP.new(
+        "source" => "[target][ip]",
+        "target" => "target",
+        "fields" => [ "city_name", "region_name" ],
+        "add_tag" => "done", "database" => CITYDB
+      )
+    }
+
+    before do
+      plugin.register
+      plugin.filter(event)
+    end
+
+    context "when source field 'ip' is a subfield of 'target'" do
+
+      it "should preserve value in [target][ip]" do
+        expect(event.get("[target][ip]")).to eq("173.9.34.107")
+      end
+
+      it "should set other subfields of 'target' properly" do
+        expect(event.get("target").to_hash.keys.sort).to eq(["city_name", "ip", "region_name"])
+        expect(event.get("[target][city_name]")).to eq("Mendon")
+        expect(event.get("[target][region_name]")).to eq("Massachusetts")
+      end
+
+    end
+
   end
 
   describe "correct encodings with default db" do
@@ -104,14 +115,14 @@ describe LogStash::Filters::GeoIP do
     CONFIG
     expected_fields = %w(ip country_code2 country_code3 country_name
                            continent_code region_name city_name postal_code
-                           dma_code area_code timezone)
+                           dma_code timezone)
 
     sample("ip" => "1.1.1.1") do
       checked = 0
       expected_fields.each do |f|
-        next unless subject["geoip"][f]
+        next unless subject.get("geoip")[f]
         checked += 1
-        insist { subject["geoip"][f].encoding } == Encoding::UTF_8
+        insist { subject.get("geoip")[f].encoding } == Encoding::UTF_8
       end
       insist { checked } > 0
     end
@@ -119,38 +130,13 @@ describe LogStash::Filters::GeoIP do
     sample("ip" => "189.2.0.0") do
       checked = 0
       expected_fields.each do |f|
-        next unless subject["geoip"][f]
+        next unless subject.get("geoip")[f]
         checked += 1
-        insist { subject["geoip"][f].encoding } == Encoding::UTF_8
+        insist { subject.get("geoip")[f].encoding } == Encoding::UTF_8
       end
       insist { checked } > 0
     end
 
-  end
-
-  describe "correct encodings with ASN db" do
-    config <<-CONFIG
-      filter {
-        geoip {
-          source => "ip"
-          database => "#{ASNDB}"
-        }
-      }
-    CONFIG
-
-
-    sample("ip" => "1.1.1.1") do
-      insist { subject["geoip"]["asn"].encoding } == Encoding::UTF_8
-    end
-    sample("ip" => "187.2.0.0") do
-      insist { subject["geoip"]["asn"].encoding } == Encoding::UTF_8
-    end
-    sample("ip" => "189.2.0.0") do
-      insist { subject["geoip"]["asn"].encoding } == Encoding::UTF_8
-    end
-    sample("ip" => "161.24.0.0") do
-      insist { subject["geoip"]["asn"].encoding } == Encoding::UTF_8
-    end
   end
 
   describe "location field" do
@@ -164,7 +150,7 @@ describe LogStash::Filters::GeoIP do
       end
 
       it "should have a location field" do
-        expect(event["[geoip][location]"]).not_to(be_nil)
+        expect(event.get("[geoip][location]")).not_to(be_nil)
       end
     end
 
@@ -189,8 +175,9 @@ describe LogStash::Filters::GeoIP do
           filter {
             geoip {
               source => "ip"
-              database => "#{ASNDB}"
-              filter_private_ips => false
+
+              database => "#{CITYDB}"
+
             }
           }
         CONFIG
@@ -205,7 +192,9 @@ describe LogStash::Filters::GeoIP do
     end
 
     describe "filter method outcomes" do
-      let(:plugin) { LogStash::Filters::GeoIP.new("source" => "message", "add_tag" => "done", "database" => ASNDB, "filter_private_ips" => false) }
+
+      let(:plugin) { LogStash::Filters::GeoIP.new("source" => "message", "add_tag" => "done", "database" => CITYDB) }
+
       let(:event) { LogStash::Event.new("message" => ipstring) }
 
       before do
@@ -218,11 +207,11 @@ describe LogStash::Filters::GeoIP do
         let(:ipstring) { "N/A" }
 
         it "should set the target field to an empty hash" do
-          expect(event["geoip"]).to eq({})
+          expect(event.get("geoip")).to eq({})
         end
 
-        it "should not have added any tags" do
-          expect(event["tags"]).to be_nil
+        it "should add failure tags" do
+          expect(event.get("tags")).to include("_geoip_lookup_failure")
         end
       end
 
@@ -231,9 +220,39 @@ describe LogStash::Filters::GeoIP do
         let(:ipstring) { "123.45.67.89,61.160.232.222" }
         
         it "should set the target field to an empty hash" do
-          expect(event["geoip"]).to eq({})
+          expect(event.get("geoip")).to eq({})
         end
       end
+
+      context "when a IP is not found in the DB" do
+        let(:ipstring) { "0.0.0.0" }
+
+        it "should set the target field to an empty hash" do
+          expect(event.get("geoip")).to eq({})
+          expect(event.get("tags")).to include("_geoip_lookup_failure")
+        end
+      end
+
+      context "when IP is IPv6 format for localhost" do
+        let(:ipstring) { "::1" }
+
+        it "should set the target field to an empty hash" do
+          expect(event.get("geoip")).to eq({})
+        end
+      end
+
+      context "when IP is valid IPv6 format" do
+        let(:ipstring) { "2607:f0d0:1002:51::4" }
+
+        it "should set the target fields properly" do
+          expect(event.get("geoip")).not_to be_empty
+          expect(event.get("geoip")["ip"]).to eq("2607:f0d0:1002:51:0:0:0:4")
+          expect(event.get("geoip").to_hash.keys.sort).to eq(
+            ["continent_code", "country_code2", "country_code3", "country_name", "ip", "latitude", "location", "longitude"]
+          )
+        end
+      end
+
     end
 
     context "should return the correct source field in the logging message" do
@@ -245,94 +264,7 @@ describe LogStash::Filters::GeoIP do
 
   end
 
-  describe "multiple ips" do
-    config <<-CONFIG
-          filter {
-            geoip {
-              source => "ip"
-              database => "#{ASNDB}"
-              filter_private_ips => true
-            }
-          }
-        CONFIG
-
-    describe "filter method outcomes" do
-      let(:plugin) { LogStash::Filters::GeoIP.new("source" => "message", "add_tag" => "done", "database" => ASNDB) }
-      let(:event) { LogStash::Event.new("message" => ipstring) }
-
-      before do
-        plugin.register
-        plugin.filter(event)
-      end
-
-      context "when multiple ips are found" do
-        # regression test for issue https://github.com/logstash-plugins/logstash-filter-geoip/issues/51
-        let(:ipstring) { "123.45.67.89,61.160.232.222" }
-
-        it "should take the first public ip" do # {"number"=>"AS6619", "asn"=>"SamsungSDS Inc."}
-          expect(event["geoip"]["number"]).not_to be_nil
-          expect(event["geoip"]["asn"]).not_to be_nil
-        end
-      end
-    end
-
-  end 
-
-  describe "private ips" do
-    let(:plugin) { LogStash::Filters::GeoIP.new("source" => "message", "add_tag" => "done", "database" => ASNDB) }
-    before do
-      plugin.register
-    end
-    
-    context "when ip is in a private 10/8 network" do
-
-      let(:event) { LogStash::Event.new("message" => "10.1.2.3" ) }
-      before do
-        plugin.filter(event)
-      end
-
-      it "should return nil" do
-        expect(event["geoip"]).to be_nil
-        end # it
-      end # context
-
-     context "when ip is in a private 172.16/12 network" do
-
-      let(:event) { LogStash::Event.new("message" => "172.16.0.8" ) }
-      before do
-        plugin.filter(event)
-      end
-
-      it "should return nil" do
-        expect(event["geoip"]).to be_nil
-        end # it
-      end # context
-
-     context "when ip is in a private 192.168/16 network" do
-
-      let(:event) { LogStash::Event.new("message" => "192.168.8.15" ) }
-      before do
-        plugin.filter(event)
-      end
-
-      it "should return nil" do
-        expect(event["geoip"]).to be_nil
-        end # it
-      end # context
-
-     context "when ip is not in a private network" do
-
-      let(:event) { LogStash::Event.new("message" => "172.15.0.8" ) }
-      before do
-        plugin.filter(event)
-      end
-
-      it "should not return nil" do
-        expect(event["geoip"]).not_to be_nil
-        end # it
-      end # context
-      
-    end # describe
+  
 
 
 
@@ -350,46 +282,9 @@ describe LogStash::Filters::GeoIP do
     context "should return the correct sourcefield in the logging message" do
       sample("ip" => "8.8.8.8") do
         expect(LogStash::Filters::GeoIP.logger).to receive(:error).with(anything, include(:field => "ip"))
-        subject
+        expect { subject }.to raise_error
       end
     end
   end
 
-  describe "returned object identities" do
-    let(:plugin) { LogStash::Filters::GeoIP.new("source" => "message") }
-    let(:event) { LogStash::Event.new("message" => "8.8.8.8") }
-    let(:alt_event) { LogStash::Event.new("message" => "8.8.8.8") }
-
-    before do
-      plugin.register
-    end
-
-    it "should dup the objects" do
-      plugin.apply_geodata(plugin.get_geo_data(event), event)
-      plugin.apply_geodata(plugin.get_geo_data(alt_event), alt_event)
-
-      event["geoip"].each do |k,v|
-        alt_v = alt_event["geoip"][k]
-        expect(v).to eql(alt_v)
-        unless v.is_a?(Numeric) # Numeric values can't be mutated, so this isn't an issue, its really for strings
-          expect(v.object_id).not_to eql(alt_v.object_id), "Object Ids for key #{k} and v #{v}"
-        end
-      end
-    end
-  end
-
-  describe "re-initializing thread current DB" do
-    let(:plugin) { LogStash::Filters::GeoIP.new("source" => "message") }
-
-    before do
-      plugin.register
-    end
-
-    it "should initialize the DB on lookup, regardless of thread state" do
-      Thread.current[plugin.threadkey] = nil
-      expect {
-        plugin.get_geo_data_for_ip("8.8.8.8")
-      }.not_to raise_error
-    end
-  end
 end